requirements.txt
Flask>=2.2
Werkzeug>=2.2
numpy
scikit-learn
tensorflow
flask
firebase-admin
flask_cors
python-dotenv